ddl: Refactor for fixing unstable drop table (#8421)
ref pingcap/tiflash#8395
