+++ dbms/src/TiDB/Schema/SchemaBuilder.cpp
@@ -39,7 +39,6 @@
 #include <Storages/KVStore/TMTContext.h>
 #include <Storages/MutableSupport.h>
 #include <TiDB/Decode/TypeMapping.h>
-#include <TiDB/Schema/SchemaBuilder-internal.h>
 #include <TiDB/Schema/SchemaBuilder.h>
 #include <TiDB/Schema/SchemaNameMapper.h>
 #include <TiDB/Schema/TiDB.h>
@@ -58,128 +57,6 @@ namespace ErrorCodes
 extern const int DDL_ERROR;
 extern const int SYNTAX_ERROR;
 } // namespace ErrorCodes
-
-void TableIDMap::doEmplaceTableID(
-    TableID table_id,
-    DatabaseID database_id,
-    std::string_view log_prefix,
-    const std::unique_lock<std::shared_mutex> &)
-{
-    if (auto iter = table_id_to_database_id.find(table_id); //
-        iter != table_id_to_database_id.end())
-    {
-        if (iter->second != database_id)
-        {
-            LOG_WARNING(
-                log,
-                "{}table_id to database_id is being overwrite, table_id={}"
-                " old_database_id={} new_database_id={}",
-                log_prefix,
-                table_id,
-                iter->second,
-                database_id);
-            iter->second = database_id;
-        }
-    }
-    else
-        table_id_to_database_id.emplace(table_id, database_id);
-}
-
-void TableIDMap::doEmplacePartitionTableID(
-    TableID partition_id,
-    TableID table_id,
-    std::string_view log_prefix,
-    const std::unique_lock<std::shared_mutex> &)
-{
-    if (auto iter = partition_id_to_logical_id.find(partition_id); //
-        iter != partition_id_to_logical_id.end())
-    {
-        if (iter->second != table_id)
-        {
-            LOG_WARNING(
-                log,
-                "{}partition_id to table_id is being overwrite, physical_table_id={}"
-                " old_logical_table_id={} new_logical_table_id={}",
-                log_prefix,
-                partition_id,
-                iter->second,
-                table_id);
-            iter->second = table_id;
-        }
-    }
-    else
-        partition_id_to_logical_id.emplace(partition_id, table_id);
-}
-
-void TableIDMap::exchangeTablePartition(
-    DatabaseID non_partition_database_id,
-    TableID non_partition_table_id,
-    DatabaseID /*partition_database_id*/,
-    TableID partition_logical_table_id,
-    TableID partition_physical_table_id)
-{
-    // Change all under the same lock
-    std::unique_lock lock(mtx_id_mapping);
-    // erase the non partition table
-    if (auto iter = table_id_to_database_id.find(non_partition_table_id); iter != table_id_to_database_id.end())
-        table_id_to_database_id.erase(iter);
-    else
-        LOG_WARNING(
-            log,
-            "ExchangeTablePartition: non partition table not in table_id_to_database_id, table_id={}",
-            non_partition_table_id);
-
-    // make the partition table to be a non-partition table
-    doEmplaceTableID(partition_physical_table_id, non_partition_database_id, "ExchangeTablePartition: ", lock);
-
-    // remove the partition table to logical table mapping
-    if (auto iter = partition_id_to_logical_id.find(partition_physical_table_id);
-        iter != partition_id_to_logical_id.end())
-    {
-        partition_id_to_logical_id.erase(iter);
-    }
-    else
-    {
-        LOG_WARNING(
-            log,
-            "ExchangeTablePartition: partition table not in partition_id_to_logical_id, physical_table_id={}",
-            partition_physical_table_id);
-    }
-
-    // make the non partition table as a partition to logical table
-    doEmplacePartitionTableID(non_partition_table_id, partition_logical_table_id, "ExchangeTablePartition: ", lock);
-}
-
-std::tuple<bool, DatabaseID, TableID> TableIDMap::findDatabaseIDAndLogicalTableID(TableID physical_table_id) const
-{
-    std::shared_lock<std::shared_mutex> lock(mtx_id_mapping);
-    DatabaseID database_id = -1;
-    if (auto database_iter = table_id_to_database_id.find(physical_table_id);
-        database_iter != table_id_to_database_id.end())
-    {
-        database_id = database_iter->second;
-        // This is a non-partition table or the logical_table of partition table.
-        return {true, database_id, physical_table_id};
-    }
-
-    /// if we can't find physical_table_id in table_id_to_database_id,
-    /// we should first try to find it in partition_id_to_logical_id because it could be the pysical_table_id of partition tables
-    TableID logical_table_id = -1;
-    if (auto logical_table_iter = partition_id_to_logical_id.find(physical_table_id);
-        logical_table_iter != partition_id_to_logical_id.end())
-    {
-        logical_table_id = logical_table_iter->second;
-        // try to get the database_id of logical_table_id
-        if (auto database_iter = table_id_to_database_id.find(logical_table_id);
-            database_iter != table_id_to_database_id.end())
-        {
-            database_id = database_iter->second;
-            // This is a non-partition table or the logical_table of partition table.
-            return {true, database_id, logical_table_id};
-        }
-    }
-    return {false, 0, 0};
-}
 
 bool isReservedDatabase(Context & context, const String & database_name)
 {
@@ -488,7 +365,10 @@ void SchemaBuilder<Getter, NameMapper>::applySetTiFlashReplica(DatabaseID databa
         auto storage = tmt_context.getStorages().get(keyspace_id, table_info->id);
         if (unlikely(storage == nullptr))
         {
-            LOG_ERROR(log, "table is not exist in TiFlash, applySetTiFlashReplica is ignored, table_id={}", table_id);
+            LOG_ERROR(
+                log,
+                "Storage instance is not exist in TiFlash, applySetTiFlashReplica is ignored, table_id={}",
+                table_id);
             return;
         }
 
@@ -581,7 +461,10 @@ void SchemaBuilder<Getter, NameMapper>::applyPartitionDiff(DatabaseID database_i
     auto storage = tmt_context.getStorages().get(keyspace_id, table_info->id);
     if (storage == nullptr)
     {
-        LOG_ERROR(log, "table is not exist in TiFlash, table_id={}", table_id);
+        LOG_ERROR(
+            log,
+            "logical_table storage instance is not exist in TiFlash, applyPartitionDiff is ignored, table_id={}",
+            table_id);
         return;
     }
 
@@ -697,7 +580,10 @@ void SchemaBuilder<Getter, NameMapper>::applyRenameTable(DatabaseID database_id,
     auto storage = tmt_context.getStorages().get(keyspace_id, table_id);
     if (storage == nullptr)
     {
-        LOG_WARNING(log, "table is not exist in TiFlash, applyRenameTable is ignored, table_id={}", table_id);
+        LOG_WARNING(
+            log,
+            "Storage instance is not exist in TiFlash, applyRenameTable is ignored, table_id={}",
+            table_id);
         return;
     }
 
@@ -722,7 +608,8 @@ void SchemaBuilder<Getter, NameMapper>::applyRenameLogicalTable(
             {
                 LOG_ERROR(
                     log,
-                    "table is not exist in TiFlash, applyRenamePhysicalTable is ignored, physical_table_id={}",
+                    "Storage instance is not exist in TiFlash, applyRenamePhysicalTable is ignored, "
+                    "physical_table_id={}",
                     part_def.id);
                 return;
             }
@@ -942,10 +829,7 @@ void SchemaBuilder<Getter, NameMapper>::applyCreateSchema(const TiDB::DBInfoPtr 
     interpreter.setForceRestoreData(false);
     interpreter.execute();
 
-    {
-        std::unique_lock<std::shared_mutex> lock(shared_mutex_for_databases);
-        databases.emplace(db_info->id, db_info);
-    }
+    databases.addDatabaseInfo(db_info);
 
     LOG_INFO(log, "Create database {} end, database_id={}", name_mapper.debugDatabaseName(*db_info), db_info->id);
 }
@@ -953,16 +837,11 @@ void SchemaBuilder<Getter, NameMapper>::applyCreateSchema(const TiDB::DBInfoPtr 
 template <typename Getter, typename NameMapper>
 void SchemaBuilder<Getter, NameMapper>::applyDropSchema(DatabaseID schema_id)
 {
-    TiDB::DBInfoPtr db_info;
+    TiDB::DBInfoPtr db_info = databases.getDBInfo(schema_id);
+    if (unlikely(db_info == nullptr))
     {
-        std::shared_lock<std::shared_mutex> shared_lock(shared_mutex_for_databases);
-        auto it = databases.find(schema_id);
-        if (unlikely(it == databases.end()))
-        {
-            LOG_INFO(log, "Try to drop database but not found, may has been dropped, database_id={}", schema_id);
-            return;
-        }
-        db_info = it->second;
+        LOG_INFO(log, "Try to drop database but not found, may has been dropped, database_id={}", schema_id);
+        return;
     }
 
     {
@@ -974,10 +853,7 @@ void SchemaBuilder<Getter, NameMapper>::applyDropSchema(DatabaseID schema_id)
 
     applyDropSchema(name_mapper.mapDatabaseName(*db_info));
 
-    {
-        std::unique_lock<std::shared_mutex> lock(shared_mutex_for_databases);
-        databases.erase(schema_id);
-    }
+    databases.eraseDBInfo(schema_id);
 }
 
 template <typename Getter, typename NameMapper>
@@ -1286,24 +1162,24 @@ void SchemaBuilder<Getter, NameMapper>::syncAllSchema()
     for (const auto & db : all_schemas)
     {
         auto task = [this, db, &created_db_set, &created_db_set_mutex] {
+            do
             {
-                std::shared_lock<std::shared_mutex> shared_lock(shared_mutex_for_databases);
-                if (databases.find(db->id) == databases.end())
+                if (databases.exists(db->id))
                 {
-                    shared_lock.unlock();
-                    applyCreateSchema(db);
-                    {
-                        std::unique_lock<std::mutex> created_db_set_lock(created_db_set_mutex);
-                        created_db_set.emplace(name_mapper.mapDatabaseName(*db));
-                    }
-
-                    LOG_INFO(
-                        log,
-                        "Database {} created during sync all schemas, database_id={}",
-                        name_mapper.debugDatabaseName(*db),
-                        db->id);
+                    break;
                 }
-            }
+                applyCreateSchema(db);
+                {
+                    std::unique_lock<std::mutex> created_db_set_lock(created_db_set_mutex);
+                    created_db_set.emplace(name_mapper.mapDatabaseName(*db));
+                }
+
+                LOG_INFO(
+                    log,
+                    "Database {} created during sync all schemas, database_id={}",
+                    name_mapper.debugDatabaseName(*db),
+                    db->id);
+            } while (false); // Ensure database existing
 
             std::vector<TableInfoPtr> tables = getter.listTables(db->id);
             for (auto & table : tables)
