+++ dbms/src/TiDB/Schema/SchemaSyncService.cpp
@@ -258,6 +258,8 @@ bool SchemaSyncService::gc(Timestamp gc_safepoint, KeyspaceID keyspace_id)
         }
     }
 
+    auto schema_sync_manager = tmt_context.getSchemaSyncerManager();
+
     // Physically drop tables
     bool succeeded = true;
     for (auto & storage_ptr : storages_to_gc)
@@ -271,25 +273,22 @@ bool SchemaSyncService::gc(Timestamp gc_safepoint, KeyspaceID keyspace_id)
         String table_name = storage->getTableName();
         const auto & table_info = storage->getTableInfo();
 
-        tmt_context.getSchemaSyncerManager()->removeTableID(keyspace_id, table_info.id);
-
         auto canonical_name = [&]() {
-            // DB info maintenance is parallel with GC logic so we can't always assume one specific DB info's existence, thus checking its validity.
-            auto db_info = tmt_context.getSchemaSyncerManager()->getDBInfoByMappedName(keyspace_id, database_name);
-            return db_info ? fmt::format(
-                       "{}, database_id={} table_id={}",
-                       SchemaNameMapper().debugCanonicalName(*db_info, table_info),
-                       db_info->id,
-                       table_info.id)
-                           : fmt::format(
-                               "({}).{}, table_id={}",
-                               database_name,
-                               SchemaNameMapper().debugTableName(table_info),
-                               table_info.id);
+            auto database_id = SchemaNameMapper::tryGetDatabaseID(database_name);
+            if (!database_id.has_value())
+            {
+                return fmt::format("{}.{} table_id={}", database_name, table_name, table_info.id);
+            }
+            return fmt::format(
+                "{}.{} database_id={} table_id={}",
+                database_name,
+                table_name,
+                *database_id,
+                table_info.id);
         }();
         LOG_INFO(
             keyspace_log,
-            "Physically dropping table, table_tombstone={} safepoint={} {}",
+            "Physically drop table begin, table_tombstone={} safepoint={} {}",
             storage->getTombstone(),
             gc_safepoint,
             canonical_name);
@@ -303,7 +302,9 @@ bool SchemaSyncService::gc(Timestamp gc_safepoint, KeyspaceID keyspace_id)
         {
             InterpreterDropQuery drop_interpreter(ast_drop_query, context);
             drop_interpreter.execute();
-            LOG_INFO(keyspace_log, "Physically dropped table {}", canonical_name);
+            LOG_INFO(keyspace_log, "Physically drop table {} end", canonical_name);
+            // remove the id mapping after physically dropped
+            schema_sync_manager->removeTableID(keyspace_id, table_info.id);
             ++num_tables_removed;
         }
         catch (DB::Exception & e)
@@ -347,7 +348,7 @@ bool SchemaSyncService::gc(Timestamp gc_safepoint, KeyspaceID keyspace_id)
             continue;
         }
 
-        LOG_INFO(keyspace_log, "Physically dropping database, database_tombstone={} {}", db->getTombstone(), db_name);
+        LOG_INFO(keyspace_log, "Physically drop database begin, database_tombstone={} {}", db->getTombstone(), db_name);
         auto drop_query = std::make_shared<ASTDropQuery>();
         drop_query->database = db_name;
         drop_query->if_exists = true;
@@ -357,7 +358,7 @@ bool SchemaSyncService::gc(Timestamp gc_safepoint, KeyspaceID keyspace_id)
         {
             InterpreterDropQuery drop_interpreter(ast_drop_query, context);
             drop_interpreter.execute();
-            LOG_INFO(keyspace_log, "Physically dropped database {}, safepoint={}", db_name, gc_safepoint);
+            LOG_INFO(keyspace_log, "Physically drop database {} end, safepoint={}", db_name, gc_safepoint);
             ++num_databases_removed;
         }
         catch (DB::Exception & e)
