+++ dbms/src/TiDB/Schema/tests/gtest_name_mapper.cpp
@@ -1,0 +1,30 @@
+// Copyright 2023 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include <TiDB/Schema/SchemaNameMapper.h>
+#include <gtest/gtest.h>
+
+namespace DB::tests
+{
+
+TEST(SchemaNameMapperTest, ParseDatabaseID)
+{
+    SchemaNameMapper mapper;
+    ASSERT_EQ(10086, *mapper.tryGetDatabaseID("db_10086"));
+    ASSERT_EQ(10086, *mapper.tryGetDatabaseID("ks_100_db_10086"));
+    ASSERT_EQ(10086, *mapper.tryGetDatabaseID(mapper.mapDatabaseName(10086, 100)));
+    ASSERT_FALSE(mapper.tryGetDatabaseID("abcdefg"));
+}
+
+} // namespace DB::tests
